feat: added weapons and sets to resource manager
